+++ src/parser/enrichers/trait/_module.mjs
@@ -21,3 +21,4 @@ export * as Wulven from "./wulven/_module.mjs";
 export * as Tortle from "./tortle/_module.mjs";
 export * as Shadarkai from "./shadar-kai/_module.mjs";
 export * as Changeling from "./changeling/_module.mjs";
+export * as Khoravar from "./khoravar/_module.mjs";