+++ src/parser/enrichers/trait/khoravar/_module.mjs
@@ -1,0 +1,2 @@
+export { default as LethargyResilience } from "./LethargyResilience.mjs";
+export { default as LethargyResilienceRecharge } from "./LethargyResilienceRecharge.mjs";