+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Fixes to Haste 2024 expiring at turn end if using midi
 - Hunters Mark automation effect would expire after an hour rather than persisting.
-- Kalashtar parsing fixes
+- Kalashtar, Khoravar parsing fixes
 
 # 6.6.24